feat: support custom hex colors in calendar + extract popover hook
